# docs/apifox-full.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AuraStack 接口文档（自动生成）",
    "version": "1.0.0",
    "description": "AuraStack 接口文档（由脚本自动生成）。"
  },
  "servers": [
    {
      "url": "http://localhost:5001",
      "description": "Development"
    },
    {
      "url": "http://localhost:5000",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Flask Session Cookie 登录态"
      }
    }
  },
  "paths": {
...
    "/api/admin/logs/login": {
      "get": {
        "tags": [
          "后台-日志管理"
        ],
        "summary": "查询login",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/日志管理",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/logs/operation": {
      "get": {
        "tags": [
          "后台-日志管理"
        ],
        "summary": "查询operation",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/日志管理",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/api/admin/logs/login/export": {
      "post": {
        "tags": [
          "后台-日志管理"
        ],
        "summary": "导出login_logs",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/日志管理",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/admin/logs/login/template": {
      "get": {
        "tags": [
          "后台-日志管理"
        ],
        "summary": "下载login_logs导入模板",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/日志管理",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/logs/login/import": {
      "post": {
        "tags": [
          "后台-日志管理"
        ],
        "summary": "导入login_logs",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/日志管理",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/admin/logs/operation/export": {
      "post": {
        "tags": [
          "后台-日志管理"
        ],
        "summary": "导出operation_logs",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/日志管理",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/admin/logs/operation/template": {
      "get": {
        "tags": [
          "后台-日志管理"
        ],
        "summary": "下载operation_logs导入模板",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/日志管理",
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/api/admin/logs/operation/import": {
      "post": {
        "tags": [
          "后台-日志管理"
        ],
        "summary": "导入operation_logs",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误"
          },
          "401": {
            "description": "未授权"
          },
          "403": {
            "description": "无权限"
          },
          "500": {
            "description": "服务器错误"
          }
        },
        "x-apifox-folder": "后台/系统管理/日志管理",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
  }
}